Update customer.php
wrong  format type for currency

+++ upload/admin/model/customer/customer.php
@@ -334,10 +334,10 @@ class Customer extends \Opencart\System\Engine\Model {
 		return (int)$query->row['total'];
 	}
 
-	public function getTransactionTotal(int $customer_id): int {
+	public function getTransactionTotal(int $customer_id): float {
 		$query = $this->db->query("SELECT SUM(`amount`) AS `total` FROM `" . DB_PREFIX . "customer_transaction` WHERE `customer_id` = '" . (int)$customer_id . "'");
 
-		return (int)$query->row['total'];
+		return (float)$query->row['total'];
 	}
 
 	public function getTotalTransactionsByOrderId(int $order_id): int {
